完成数据权限模块,根据mybatis-plus和jsqlParser实现

+++ lfl-module-system/src/main/java/cn/iocoder/lfl/module/system/controller/admin/user/UserController.java
@@ -84,7 +84,7 @@ public class UserController {
     @GetMapping("/get")
     @Operation(summary = "获得用户详情")
     @Parameter(name = "id", description = "用户编号", required = true, example = "1024")
-    public CommonResult<UserRespVO> getUser(@RequestParam("id") Integer id){
+    public CommonResult<UserRespVO> getUser(@RequestParam("id") Long id){
         UserRespVO user = userService.getUser(id);
         return CommonResult.success(user);
     }
