+++ lfl-module-system/src/main/java/cn/iocoder/lfl/module/system/service/user/AdminUserService.java
@@ -20,7 +20,7 @@ public interface AdminUserService {
 
      void updateUserLogin(Long id,String loginIp);
 
-     UserRespVO getUser(Integer id);
+     UserRespVO getUser(Long id);
 
      void updateUserPassword(Long id,String password);
